+++ src/HetrickCV.cpp
@@ -53,6 +53,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelPhasorReset);
 	p->addModel(modelPhasorShape);
 	p->addModel(modelPhasorSubstepShape);
+	p->addModel(modelPhasorTimetable);
 	p->addModel(modelPhasorToLFO);
 	p->addModel(modelPhasorToWaveforms);
 	p->addModel(modelProbability);